refactor: isolate exceptions and errors
