+++ lib/uploadcare/client/rest_client.rb
@@ -5,10 +5,12 @@ require 'uploadcare/concerns/throttle_handler'
 
 module Uploadcare
   module Client
+    # @abstract
     # General client for signed REST requests
     class RestClient < ApiStruct::Client
-      include Uploadcare::ErrorHandler
-      include Uploadcare::ThrottleHandler
+      include Uploadcare::Concerns::ErrorHandler
+      include Uploadcare::Concerns::ThrottleHandler
+      include Exception
       rest_api 'files'
 
       alias api_struct_delete delete
@@ -20,10 +22,8 @@ module Uploadcare
 
       def request(method: 'GET', uri:, **options)
         headers = AuthenticationHeader.call(method: method.upcase, uri: uri, **options)
-        handle_throttling do
-          send('api_struct_' + method.downcase, path: remove_trailing_slash(uri),
-                                                headers: headers, body: options[:content])
-        end
+        handle_throttling { send('api_struct_' + method.downcase, path: remove_trailing_slash(uri),
+           headers: headers, body: options[:content]) }
       end
 
       def get(**options)
