+++ lib/uploadcare/concerns/upload_error_handler.rb
@@ -1,24 +1,28 @@
 # frozen_string_literal: true
 
 module Uploadcare
-  # Wrapper for responses
-  # raises errors instead of returning monads
-  module UploadErrorHandler
-    def failure(response)
-      catch_throttling_error(response)
-      parsed_response = JSON.parse(response.body.to_s)
-      raise RequestError, parsed_response['detail']
-    rescue JSON::ParserError
-      raise RequestError, response.status
-    end
+  module Concerns
+    # Wrapper for responses
+    # raises errors instead of returning monads
+    module UploadErrorHandler
+      include Exception
 
-    private
+      def failure(response)
+        catch_throttling_error(response)
+        parsed_response = JSON.parse(response.body.to_s)
+        raise RequestError, parsed_response['detail']
+      rescue JSON::ParserError
+        raise RequestError, response.status
+      end
 
-    def catch_throttling_error(response)
-      return unless response.code == 429
+      private
 
-      retry_after = response.headers['Retry-After'].to_i + 1 || 11
-      raise ThrottleError.new(retry_after), "Response throttled, retry #{retry_after} seconds later"
+      def catch_throttling_error(response)
+        return unless response.code == 429
+
+        retry_after = response.headers['Retry-After'].to_i + 1 || 11
+        raise ThrottleError.new(retry_after), "Response throttled, retry #{retry_after} seconds later"
+      end
     end
   end
 end
