+++ lib/uploadcare/exception/request_error.rb
@@ -1,0 +1,9 @@
+# frozen_string_literal: true
+
+module Uploadcare
+  module Exception
+    # Standard error for invalid API responses
+    class RequestError < StandardError
+    end
+  end
+end
