+++ spec/spec_helper.rb
@@ -8,6 +8,7 @@ require 'uploadcare'
 Dir[File.expand_path(File.join(File.dirname(__FILE__),'support','**','*.rb'))].each { |f| require f }
 
 RSpec.configure do |config|
+  include Uploadcare::Exception
   # Enable flags like --only-failures and --next-failure
   config.example_status_persistence_file_path = ".rspec_status"
